+++ WeatherForecast/WeatherForecast/Extension/StringExtension.swift
@@ -1,0 +1,10 @@
+//
+//  StringExtension.swift
+//  WeatherForecast
+//
+//  Created by 리나 on 2021/01/23.
+//
+
+extension String {
+    static var empty = ""
+}
